Cleanup, remove sparse_thread.cpp.
